Add files via upload
Team Members: Kaustubh Nigam and Rahul Agarwal

UID: 3035825112 and 3035756555

Name of the Game: The Stone of Immortality

Description

An interactive text-based fantasy game that takes the player on an adventurous journey filled with monsters, dragons and wizards. The player will have to answer questions and battle opponents which will help them advance on this journey to find the ultimate treasure – The Stone of Immortality. The game will be divided into various levels with a surprise pop-up quiz at any random level that would lead to rewards. How will coding requirements support the game features?

The pop-up quiz will be generated at any random level(any random point in the game). The enemy for the given level will be decided randomly.
An array stores the names of the enemies defeated and writes it out in the output file.
Weapon input incorporates dynamic memory mananagement.
The statistics of the game are stored in an output file. Weapon names and opponent names are stored in separate text files which are used in the program.
The game is divided into multiple files(Header files, cpp file, text files)
There is proper indentation and naming styles.
There is in-code documentation